refactor: add dot in import
