+++ workflows/activity_diagram_workflow/workflow.py
@@ -10,7 +10,7 @@ from typing import TypedDict, Optional, List
 from workflows.nodes import get_chat_history, get_content_file
 from connect_model import get_model_client, MODEL
 from ..utils import extractor
-from ..response import success_response, error_response
+from ...response import success_response, error_response
 
 # from services.mermaid_validator.subprocess_manager import MermaidSubprocessManager
