+++ workflows/high_level_requirements_workflow/workflow.py
@@ -8,7 +8,7 @@ from typing import TypedDict, Optional, List
 from workflows.nodes import get_chat_history, get_content_file
 from connect_model import get_model_client, MODEL
 from ..utils import extractor
-from ..response import success_response, error_response
+from ...response import success_response, error_response
 
 class HighLevelRequirementsState(TypedDict):
     user_message: str
