+++ workflows/lld_arch_workflow/workflow.py
@@ -11,7 +11,7 @@ from workflows.nodes import get_chat_history, get_content_file
 from connect_model import get_model_client, MODEL
 from models.lld_arch import LLDArchResponse, LLDArchOutput
 from ..utils import extractor
-from ..response import success_response, error_response
+from ...response import success_response, error_response
 
 logger = logging.getLogger(__name__)
