+++ workflows/lld_pseudo_workflow/workflow.py
@@ -12,7 +12,7 @@ from connect_model import get_model_client, MODEL
 from models.lld_pseudo import LLDPseudoResponse, LLDPseudoOutput
 import re
 from ..utils import extractor
-from ..response import success_response, error_response
+from ...response import success_response, error_response
 
 logger = logging.getLogger(__name__)
